requirements.txt
openai
python-dotenv
fastapi
uvicorn
pydantic-ai 
python-multipart
langchain
langgraph
langchain-openai
langchain-community
pypdf
pinecone
tiktoken
sqlalchemy
